Create function to determine correct guess outcome
The `determine_positions` function returns an ascii
representation of numbers that are either in the
cypher, or in the cypher and in the correct position.

The return value is an array that is trimmed of blanks
(wrong guesses) and shuffled so the guesser only gets
the information of whether a number they've guessed is
in the cypher and in the correct position.

The `random_number_generator` returns an array
(default length 4) of randomly chosen numbers to
attempt to match.

The file `mastermind.md` contains notes that I've
made in planning and preparation for the implementation
of the game functionality, and this is its initial
commit.