Create methods to set up configuration for game
Prompt player to set things like length of cypher, amount
of turns, are blanks and duplicates allowed into the cypher.

It's at this point that the project started naturally seeking
objects to help organize the code.  I broke the RNG code out
into its own object that is coming in a seperate commit.