Push changes for tracking
Currently things are not working as I intend them to.  I have,
through hours of research and careful consideration decided to
take @Josh's advice from discord and go with a different solver
solution for the problem.  I have implemented a reduction algo
that successfully reduces the guess pool as intended, however
it doesn't seem to be matching and returning the solution that
it should be.

Going through it with pry-byebug, I'm seeing that on the second
iteration through, randomly choosing a second guess to match against
appears to be causing it to dump the answer out of the pool.

I have more debugging to do, and as it stands, this code does not
function correctly yet.