Create guess method
Guess method handles the logic of obtaining a solve through creating
a solver object, then sending its initial guess into the feedback call.
Once it has the initial feedback from that call against the set cypher,
it calls the solver to reduce the solution pool and give feedback until
it reaches a single length solution which should match the cypher.

This method returns a two (2) length array containing a stringified
version of the solution as well as the number of turns that it took
to reach this solution.

It doesn't always get the answer in under 5 steps, but most of the
test runs I've put it through it seems to hit between 4 and 6 guesses
fairly consistantly.