Refactor get_feedback
We were getting weird feedback from the original method
as it was not accounting for positions correctly, so it would
count the same value (when given duplicates) as being both in the
correct position and in the wrong position, even if the code
it was checking against only had a single equivalent value and it
was in the correct position.

The updated method seems to work.